ListView management
Made it so that the user can't select another item from the list view
when confirm popup ups are present. Without it the user can select
several items and the popup doesn't refresh but stacks on the previous
ones. I also re-organized some code is placed in functions (check
FavoritesActivity).